Separated functions in multiple files
